bring back test for nominal config ranker + check distances
