+++ tests/distance/CMakeLists.txt
@@ -27,3 +27,13 @@ aikido_add_test(test_DistanceMetricDefaults test_DistanceMetricDefaults.cpp)
 target_link_libraries(test_DistanceMetricDefaults
   "${PROJECT_NAME}_distance"
   "${PROJECT_NAME}_statespace")
+
+aikido_add_test(test_RandomConfigurationRanker
+  test_RandomConfigurationRanker.cpp)
+target_link_libraries(test_RandomConfigurationRanker
+    "${PROJECT_NAME}_distance")
+
+aikido_add_test(test_NominalConfigurationRanker
+  test_NominalConfigurationRanker.cpp)
+target_link_libraries(test_NominalConfigurationRanker
+  "${PROJECT_NAME}_distance")
